Move VirtualDrive to a dedicated file
